% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{create_efts}
\alias{create_efts}
\title{Creates a EftsDataSet for write access to a netCDF EFTS data set}
\usage{
create_efts(fname, time_dim_info, data_var_definitions, stations_varnames,
  station_names, nc_attributes = list(), lead_length = 48,
  ensemble_length = 50)
}
\arguments{
\item{fname}{file name to create to. The file must not exist already.}

\item{time_dim_info}{a list with the units and values defining the time dimension of the data set}

\item{data_var_definitions}{a data frame, acceptable by \code{\link{create_variable_definitions}}, or list of netcdf variable definitions, e.g. 
\code{list(rain_fcast_ens=list(name='rain_fcast_ens', longname='ECMWF Rainfall ensemble forecasts', units='mm', missval=-9999.0, precision='double', attributes=list(type=2, type_description='accumulated over the preceding interval')))}}

\item{stations_varnames}{station identifiers, coercible to an integer vector (note: may change to be a more flexible character storage)}

\item{station_names}{optional; names of the stations}

\item{nc_attributes}{a named list of characters, attributes for the whole file.}

\item{lead_length}{length of the lead forecasting time series.}

\item{ensemble_length}{number of ensembles, i.e. number of forecasts for each point on the main time axis of the data set}
}
\value{
A EftsDataSet object
}
\description{
Creates a EftsDataSet for write access to a netCDF EFTS data set
}
\examples{
\dontrun{
fname <- '~/tmp/f.nc'
stations_varnames <- c(123,456)
nEns <- 3
nLead <- 4
nTimeSteps <- 12

timeAxisStart <- ISOdate(year=2010, month=08, day=01, hour = 14, min = 0, sec = 0, tz = 'UTC')
time_dim_info <- create_time_info(from=timeAxisStart, n=nTimeSteps)

# It is possible to define variables for three combinations of dimensions.
# dimensions '4' ==> [lead_time,station,ens_member,time]
# dimensions '3' ==> [station,ens_member,time]   
# dimensions '2' ==> [station,time]   

variable_names <- c('var1_fcast_ens','var2_fcast_ens', 'var1_obs', 
  'var2_obs', 'var1_ens','var2_ens')
varDef <- create_variable_definition_dataframe(variable_names=variable_names, 
  longNames = paste(variable_names, 'synthetic data'))
varDef$dimensions <- c('4','4','2','2','3','3')

snc <- create_efts(fname, time_dim_info, varDef, stations_varnames, 
  nc_attributes=create_var_attribute_definition(), lead_length=nLead, ensemble_length=nEns)

# Following is code that was used to create unit tests for EFTS.
# This is kept in this example to provide sample on now to write data of various dimension.
td <- snc$get_time_dim()
m <- matrix(ncol=nEns, nrow=nLead)
for (rnum in 1:nLead) {
    for (cnum in 1:nEns) {
      m[rnum,cnum] = rnum*0.01 + cnum*0.1
  }
}
#      [,1] [,2] [,3]
# [1,] 0.11 0.21 0.31
# [2,] 0.12 0.22 0.32
# [3,] 0.13 0.23 0.33
# [4,] 0.14 0.24 0.34
for (i in 1:length(td)) {
  for (j in 1:length(stations_varnames)) {
    station <- stations_varnames[j]
    var1Values <- i + 0.1*j + m
    var2Values <- 2*var1Values
    dtime = td[i]
    snc$put_ensemble_forecasts(var1Values, variable_name = variable_names[1], 
       identifier = station, start_time = dtime)
    snc$put_ensemble_forecasts(var2Values, variable_name = variable_names[2], 
       identifier = station, start_time = dtime)
  }
}

timeSteps <- 1:length(td)
for (j in 1:length(stations_varnames)) {
  var3Values <- timeSteps + 0.1*j
  var4Values <- var3Values + 0.01*timeSteps + 0.001*j

  station <- stations_varnames[j]
  snc$put_single_series(var3Values, variable_name = variable_names[3], identifier = station)
  snc$put_single_series(var4Values, variable_name = variable_names[4], identifier = station)
}

for (j in 1:length(stations_varnames)) {

  var5Xts <- matrix(rep(1:nEns, each=nTimeSteps) + timeSteps + 0.1*j, ncol=nEns)

  # [time,ens_member] to [ens_member,time], as expected by put_ensemble_series
  var5Values <- t(var5Xts) 
  var6Values <- 0.25 * var5Values

  station <- stations_varnames[j]
  snc$put_ensemble_series(var5Values, variable_name = variable_names[5], identifier = station)
  snc$put_ensemble_series(var6Values, variable_name = variable_names[6], identifier = station)
}
snc$close()
}
}
