% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{create_time_info}
\alias{create_time_info}
\title{Helper function to create the definition of the time dimension for use in a netCDF file}
\usage{
create_time_info(from, n, tStep = "hours since", tStepDelta = 1L, tzoffset)
}
\arguments{
\item{from}{the start date of the time axis}

\item{n}{length of the time dimension}

\item{tStep}{unit prefix in the time dimension units}

\item{tStepDelta}{integer, length of time units between each steps}

\item{tzoffset}{an optional character, the time offset from UTC, e.g. '+1000' for 10 hours ahead of UTC. Can be missing, in which case 'from' must be explicitly a UTC time. Note that the tzoffset completely supersedes the time zone if present.}
}
\value{
A list with keys units and values
}
\description{
Helper function to create the definition of the time dimension for use in a netCDF file. Defaults to create an hourly axis.
}
\seealso{
See
   \code{\link{create_efts}} for examples
}
