% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{create_variable_definition}
\alias{create_variable_definition}
\title{Create a variable definition}
\usage{
create_variable_definition(name, longname = "", units = "mm",
  missval = -9999, precision = "double", dim_type = "4",
  var_attribute = create_var_attribute_definition())
}
\arguments{
\item{name}{variable name}

\item{longname}{variable long name}

\item{units}{variable units}

\item{missval}{value code for missing data}

\item{precision}{precision}

\item{dim_type}{dimension type (EFTS integer code)}

\item{var_attribute}{list of attributes for the netCDF variable to create}
}
\value{
a list
}
\description{
Create a variable definition usable by the function \code{\link{create_efts_variables}} to create netCDF variables.
}
\examples{
\dontrun{
create_variable_definition(name='rain_fcast_ens', 
}
}
