% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{default_optional_variable_definitions_v2_0}
\alias{default_optional_variable_definitions_v2_0}
\title{Provide a template definition of optional geolocation variables}
\usage{
default_optional_variable_definitions_v2_0()
}
\value{
a data frame
}
\description{
Provide a template definition of optional geolocation and geographic variables x, y, area and elevation. 
See \url{https://github.com/csiro-hydroinformatics/efts/blob/107c553045a37e6ef36b2eababf6a299e7883d50/docs/netcdf_for_water_forecasting.md#optional-variables}.
}
\seealso{
See
   \code{\link{create_variable_definition}} and \code{\link{create_efts}} for examples
}
