% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efts_package.R
\docType{package}
\name{efts-package}
\alias{efts-package}
\alias{efts}
\title{Accessing ensemble forecast time series (EFTS) data set stored in netCDF formats}
\description{
Accessing ensemble forecast time series (EFTS) data set stored in netCDF formats
}
\details{
\tabular{ll}{
Package: \tab efts\cr
Type: \tab Package\cr
Version: \tab 0.8-0\cr
Date: \tab 2018-04-08\cr
Release Notes: \tab Improve and test subsetting and reordering of multidimensional arrays retrieved via ncdf4 \cr
License: \tab MIT \cr
}

Accessing ensemble forecast time series (EFTS) data set stored in netCDF formats, without the need for lower-level ncdf4 operations.
See \code{\link{open_efts}} \code{\link{create_efts}} for code examples to read/write files using this package.

\tabular{lll}{
Version \tab Date \tab Notes \cr
0.7.0 \tab 2018-02-14 \tab Initial commit to github \cr
0.6.x \tab 2017 \tab Non public releases of a package originally used only for unit test purposes \cr
}
}
\author{
Jean-Michel Perraud \email{jean-michel.perraud_at_csiro.au}
}
\keyword{ensemble}
\keyword{forecast}
\keyword{netCDF}
\keyword{package}
\keyword{series}
\keyword{time}
