% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{eftsdotDollarNames}
\alias{eftsdotDollarNames}
\alias{.DollarNames.EftsDataSet}
\title{method for dollar-tab-completion in R consoles.}
\usage{
\method{.DollarNames}{EftsDataSet}(x, pattern = "")
}
\arguments{
\item{x}{???}

\item{pattern}{???}
}
\description{
method for dollar-tab-completion in R consoles. 
 Credits: thanks to http://stackoverflow.com/a/26711603/2752565
}
