% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{get_time_dimension}
\alias{get_time_dimension}
\title{Retrieves the time dimension from a netcdf file}
\usage{
get_time_dimension(ncfile, time_dim_name = "time", time_zone = "UTC")
}
\arguments{
\item{ncfile}{an object of class ncdf4}

\item{time_dim_name}{The name of the time dimension, by default 'time' as per the CF conventions.}

\item{time_zone}{the time zone to use for the returned value.}
}
\value{
A vector of Dates
}
\description{
Retrieves the time dimension from a netcdf file
}
