% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_dimensions.R
\name{get_time_step_function}
\alias{get_time_step_function}
\title{Detect the unit of the time step in the time axis unit}
\usage{
get_time_step_function(time_units)
}
\arguments{
\item{time_units}{The string description of the units of the time dimension e.g. 'days since 1980-01-01' or 'hours since 2010-08-01 13:00:00 +0000'}
}
\value{
A duration function from lubridate
}
\description{
Detect the unit of the time step in the time axis unit
}
