% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_dimensions.R
\name{get_time_units}
\alias{get_time_units}
\title{Retrieves the unit string of the time dimension from a netCDF file}
\usage{
get_time_units(ncfile, time_dim_name = "time")
}
\arguments{
\item{ncfile}{an object of class ncdf4}

\item{time_dim_name}{The name of the time dimension, by default 'time' as per the CF conventions.}
}
\value{
a character
}
\description{
Retrieves the unit string of the time dimension from a netCDF file
}
