% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{NetCdfDataSet-class}
\alias{NetCdfDataSet-class}
\alias{NetCdfDataSet}
\title{Reference class convenient for access to a netCDF file.}
\description{
Reference class convenient for access to a netCDF file. Note for internal implementation that ncdf4 objects are basically lists with a class attribute.
}
\section{Fields}{

\describe{
\item{\code{ncfile}}{an object of class ncdf4}
}}

\section{Methods}{

\describe{
\item{\code{initialize(nc = NULL)}}{Create an object wrapping an ncdf4 object}
}}

