% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{open_efts}
\alias{open_efts}
\title{Creates a EftsDataSet for access to a netCDF EFTS data set}
\usage{
open_efts(ncfile, writein = FALSE)
}
\arguments{
\item{ncfile}{name of the netCDF file, or an object of class 'ncdf4'}

\item{writein}{if TRUE the data set is opened in write mode}
}
\value{
A EftsDataSet object
}
\description{
Creates a EftsDataSet for access to a netCDF EFTS data set
}
\examples{
\dontrun{
library(efts)
fname <- file.path(system.file(package='efts', 'data', 'simple.nc'))
stopifnot(file.exists(fname))
snc <- open_efts(fname)
(variable_names <- snc$get_variable_names())
stations_varnames <- snc$get_values('station_id')
nEns <- snc$get_ensemble_size()
nLead <- snc$get_lead_time_count()
td <- snc$get_time_dim()
# nTimeSteps <- 
stopifnot('runoff_ens' \%in\% variable_names)

obsRunoff <- snc$get_single_series('runoff_obs', stations_varnames[1])
plot.zoo(obsRunoff, plot.type='single')

ensRunoff <- snc$get_ensemble_series('runoff_ens', stations_varnames[1])
plot.zoo(ensRunoff, plot.type='single')

ensFcastRunoff <- snc$get_ensemble_forecasts('runoff_fcast_ens',
plot.zoo(ensFcastRunoff, plot.type='single')
}
}
