% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{create_variable_definition_dataframe}
\alias{create_variable_definition_dataframe}
\title{Create a variables definition data frame}
\usage{
create_variable_definition_dataframe(variable_names, longNames, units = "mm",
  missval = -9999, precision = "double", dimensions = "4",
  var_attribute = create_var_attribute_definition())
}
\arguments{
\item{variable_names}{character vector, names of the variables}

\item{longNames}{character vector, long names of the variables}

\item{units}{character vector}

\item{missval}{numeric vector}

\item{precision}{character vector, precision of the variables}

\item{dimensions}{character or integer vector, number of dimensions each variable (2, 3 or 4)}

\item{var_attribute}{list of attributes for the netCDF variable to create}
}
\value{
a data frame
}
\description{
Create a variable definition usable by the function \code{\link{create_variable_definitions}} 
to create netCDF variables. The use of this function is not compulsory to create a EFTS 
netCDF schema, just offered as a convenience.
}
\seealso{
See
   \code{\link{create_variable_definition}} and \code{\link{create_efts}} for examples
}
