% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{eftsdotDollarNames}
\alias{eftsdotDollarNames}
\alias{.DollarNames.EftsDataSet}
\title{method for dollar-tab-completion in R consoles.}
\usage{
\method{.DollarNames}{EftsDataSet}(x, pattern = "")
}
\arguments{
\item{x}{A reference class object}

\item{pattern}{the regex pattern to search for in potential methods. Default value is posslbly required by some versions of RStudio}
}
\description{
method for dollar-tab-completion in R consoles. It may be unnecessary in 
the future but such a method was required at some time to at least avoid 
some issues in RStudio. We may also want to customise matches compared to default reference classes.
}
